Changed to a .env.json file
