+++ cypress.config.ts
@@ -2,12 +2,9 @@ import { defineConfig } from "cypress";
 
 export default defineConfig({
   e2e: {
+    baseUrl: 'http://localhost:3000',
     setupNodeEvents(on, config) {
       // implement node event listeners here
     },
   },
-  env: {
-    gh_username: process.env.GH_USERNAME,
-    gh_password: process.env.GH_PASSWORD
-  }
 });
