+++ cypress.env.json
@@ -1,0 +1,6 @@
+{
+  "env": {
+    "gh_username": "cjodo",
+    "gh_password": "Curtispswd1!"
+  }
+}
