+++ cypress/support/e2e.ts
@@ -1,10 +1,1 @@
-import { registerCommands } from "./auth-provider-commands/github"
-registerCommands()
-
-const username = CryptoKe
-const username = Cypress.env("gh_username")
-
-beforeEach(() => {
-    cy.loginToGithub(username, Cypress.env('gh_password'))
-    cy.log("Run before specs")
-})
+require('cypress-wait-until');
