+++ cypress/support/index.ts
@@ -1,0 +1,5 @@
+declare namespace Cypress {
+    interface Chainable {
+        test(value:string): Chainable<Element>
+    }
+}
