+++ cypress/tsconfig.json
@@ -1,8 +1,10 @@
 {
+  "extends": "../tsconfig.json",
+  "include": ["**/*.ts", "../cypress.config.ts", "./types.d.ts"],
+  "exclude": [],
   "compilerOptions": {
     "target": "es5",
     "lib": ["es5", "dom"],
     "types": ["cypress", "node"]
   },
-  "include": ["**/*.ts", "../cypress.config.ts"]
 }
