+++ cypress/types.d.ts
@@ -1,6 +1,6 @@
 
 declare namespace Cypress {
-  interface Chainable<Subject = any> {
-    loginToGithub(username: string, password: string): Chainable<HTMLElement>;
+  interface Chainable {
+    loginToGithub(username: string, password: string): Chainable<any>;
   }
 }
