+++ package.json
@@ -25,6 +25,7 @@
     "@visx/scale": "^3.5.0",
     "@visx/shape": "^3.5.0",
     "@visx/tooltip": "^3.3.0",
+    "cypress-wait-until": "^3.0.1",
     "d3": "^7.9.0",
     "dotenv": "^16.4.5",
     "drizzle-orm": "^0.29.4",
